hw/mem/cxl_type3: Allow to release partial extent and extent superset in QMP interface
Before the change, the QMP interface used for add/release DC extents
only allows to release extents that exist in either pending-to-add list
or accepted list in the device, which means the DPA range of the extent must
match exactly that of an extent in either list. Otherwise, the release
request will be ignored.

With the change, we relax the constraints.
1. If the DPA range to release is within pending list, as long as the
range aligns with extents' boundaries, we allow the release.
2. If the DPA range to release fully falls in the accepted list,we allow
the release.

Signed-off-by: Fan Ni <fan.ni@samsung.com>